Dampen Magic DoT damage fixed
There was a division between UInt32 and Int32, m_totalTicks not likely
to be greater than 16 bits though.

+++ shared/game/game_unit.cpp
@@ -1146,7 +1146,7 @@ namespace wowpp
 				Int32 bonus = aura.getBasePoints();
 				if (tickCount > 1)
 				{
-					bonus /= tickCount;
+					bonus /= static_cast<Int32>(tickCount);
 				}
 
 				if (bonus < 0 && -bonus >= damage)
